🔧 [REFACTOR] renderer/types 폴더 타입 구조 정리 및 Core 타입 활용 강화
- chat-types.ts 내용을 design-types.ts로 통합하여 중복 제거
- Core 타입(ChatSessionMetadata, MessageHistory) 직접 re-export
- QuickAction, AppModeState 타입을 design-types.ts로 이동
- 하위 호환성을 위해 chat-types.ts에서 re-export 유지
- React.ComponentType<any> → Record<string, unknown>으로 타입 안전성 개선
- design 폴더 타입에서 Core 타입 직접 활용 패턴 적용
